feat: user & roles rework

+++ internal/domain/aggregates/user/user_repository.go
@@ -9,11 +9,11 @@ type Repository interface {
 	GetAll(ctx context.Context) ([]*User, error)
 	GetByEmail(ctx context.Context, email string) (*User, error)
 	GetPaginated(ctx context.Context, limit, offset int, sortBy []string) ([]*User, error)
-	GetByID(ctx context.Context, id int64) (*User, error)
+	GetByID(ctx context.Context, id uint) (*User, error)
 	Create(ctx context.Context, user *User) error
 	CreateOrUpdate(ctx context.Context, user *User) error
 	Update(ctx context.Context, user *User) error
-	UpdateLastAction(ctx context.Context, id int64) error
-	UpdateLastLogin(ctx context.Context, id int64) error
-	Delete(ctx context.Context, id int64) error
+	UpdateLastAction(ctx context.Context, id uint) error
+	UpdateLastLogin(ctx context.Context, id uint) error
+	Delete(ctx context.Context, id uint) error
 }
